requirements.txt
requests==2.28.2
RPi.GPIO==0.7.0
pyserial==3.5
